Fix nesting sections, add excludeElements & linkPrefix
Nesting sections was broken because each time it nested, currentLevelStack is emptied again and the list gets treated like the initial case, which meant that the stack gets a value pushed but the currentList does not change from the parent section, and new list items gets added at the same list level.

Fixing this meant adding a flag for evocations of buildList other than at the top level. If it is not at the top level, i.e. if it is a sublist, when the stack is empty said sublist should be created in the last list item of the previous-currentList. Also, if the sectioning element is encountered before the first heading element, any subsequent headings within the nested section are not in fact sublists; sublist flag should be set only if, at the point of encountering the sectioningElement, the stack is non-empty. Also once sublist flag is set, it persists.

Finally, we do not want every single nested section to create its own sublist; it should re-use pre-existing ones.