Only create links if id defined on heading; do not add ids
This is because the URL that is linked to is likely to persist beyond a single page load, so it would seem unwise for it to link to something that was generated on the client side on the fly, rather than something that was already canonically part of the document as served by the server.

This also conveniently sidesteps the issue of making sure that automatically generated ids do not clash with each other or with existing ids.